requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
attrs==23.1.0
crispy-bootstrap4==2022.1
Django==4.2.3
django-admin-interface==0.26.0
django-colorfield==0.9.0
django-cors-headers==4.2.0
django-crispy-forms==2.0
django-extra-fields==3.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-extra-fields==3.5.0
drf-yasg==1.21.6
filetype==1.2.0
graphene==3.2.2
graphene-django==3.1.2
graphql-core==3.2.3
graphql-relay==3.2.0
inflection==0.5.1
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
packaging==23.1
Pillow==10.0.0
promise==2.3
psycopg2==2.9.6
PyJWT==2.7.0
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0
referencing==0.29.1
rpds-py==0.8.10
six==1.16.0
sqlparse==0.4.4
swagger-spec-validator==3.0.3
text-unidecode==1.3
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
